Major player rewrite + event service implementation (#155)
* Event service framework (not working)

* Fixed 400 + report language

* Better logging + minor changes

* Working! (with a lot of improvements to make)

* Event dispatch is now asynchronous

* More clean up

* Send events properly + removed CDN_REQUEST + minor refactoring

* Compress all HTTP requests with GZip

* Fixed fields for bitrate and file size

* Send track played when shutting down or panicking

* Better handling of play origin + count decoded size

* Minor refactoring

* Rewritten player (missing crossfade) + fixed issue with cache + general refactoring

* Fixed merge issues

* Refactored instants logic

* Fixed LookupInterpolator + implemented crossfade

* Added transition metrics

* Fixed cache clean up + fixed build

* Fixed interpolator tests

* Major player rewrite done (metrics are still missing)

* Fixed crossfade not working on first track + fixed possible leaked output + minor logging changes

* Audio key metrics + preload flag + minor fixes

* Fixed build

* Removed default fade to respect original client behaviour

* Fixed crossfade issue + added logging for time offset + send metrics

* Fixed memory leak

* Fixed cache failing if first chunk is missing

* Await termination of EventService to sends all events

* Fixed repeating track + added 0x82 packet

* CDN request + remote device ID

* Fixed IllegalStateException when output has been already cleared

* Clean up

+++ core/src/main/java/xyz/gianlu/librespot/cache/JournalHeader.java
@@ -10,10 +10,10 @@ import java.util.List;
  * @author Gianlu
  */
 public final class JournalHeader {
-    public final byte id;
+    public final int id;
     public final byte[] value;
 
-    JournalHeader(byte id, @NotNull String value) {
+    JournalHeader(int id, @NotNull String value) {
         this.id = id;
         this.value = Utils.hexToBytes(value);
     }
