+++ core/src/main/java/xyz/gianlu/librespot/player/feeders/storage/AudioFileFetch.java
@@ -5,7 +5,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import xyz.gianlu.librespot.cache.CacheManager;
 import xyz.gianlu.librespot.common.Utils;
-import xyz.gianlu.librespot.player.AbsChunkedInputStream;
+import xyz.gianlu.librespot.player.feeders.AbsChunkedInputStream;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
@@ -36,7 +36,7 @@ public class AudioFileFetch implements AudioFile {
     }
 
     @Override
-    public synchronized void writeHeader(byte id, byte[] bytes, boolean cached) throws IOException {
+    public synchronized void writeHeader(int id, byte[] bytes, boolean cached) throws IOException {
         if (closed) return;
 
         if (!cached && cache != null) {
@@ -44,7 +44,7 @@ public class AudioFileFetch implements AudioFile {
                 cache.setHeader(id, bytes);
             } catch (IOException ex) {
                 if (id == HEADER_SIZE) throw new IOException(ex);
-                else LOGGER.warn(String.format("Failed writing header to cache! {id: %s}", Utils.byteToHex(id)));
+                else LOGGER.warn(String.format("Failed writing header to cache! {id: %s}", Utils.byteToHex((byte) id)));
             }
         }
